backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic>=2.4.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
aiofiles==23.2.1

# Database - Supabase / PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2

# Auth - Clerk JWT verification
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
cryptography==43.0.0

# AI Services
google-genai==0.8.0

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2

# Redis (for docker-compose)
redis[aioredis]==5.0.8

# Dev
pytest==8.3.2
pytest-asyncio==0.23.8